throw error

+++ src/postInstallHelper.js
@@ -5,14 +5,17 @@ function runPostInstall() {
   const rnPjsonFile = `${__dirname}/../../customerio-reactnative/package.json`;
   const expoPjsonFile = `${__dirname}/../package.json`;
 
-
+    // if react native SDK is installed
+    if (fs.existsSync(rnPjsonFile)) {
       const rnPJson = fs.readFileSync(rnPjsonFile, 'utf8');
 
       const rnPackage = JSON.parse(rnPJson);
       rnPackage.expoVersion = expoPjsonFile.version;
-      console.error(rnPackage, expoPjsonFile)
+
 
       fs.writeFileSync(rnPjsonFile, JSON.stringify(rnPackage, null, 2));
+      throw new Error (`RN: ${rnPJson}`);
+    }
 }
 
 exports.runPostInstall = runPostInstall;
